+++ buddycall/.kiro/specs/buddy-call/requirements.md
@@ -1,0 +1,95 @@
+# 要件定義書
+
+## はじめに
+
+Buddy Call（バディコール）は、友人、カップル、仲間などの小規模グループ（2〜5人）が相互支援と責任感を通じて個人の目標を達成することを支援するモバイルアプリケーションです。このアプリは、一人では自己管理が難しいが、仲間と一緒なら頑張れる10代〜20代のユーザーをターゲットにしています。コアコンセプトは、ユーザーが進捗を共有し、同期した集中セッションに参加し、ポジティブな相互監視を通じてモチベーションを維持できる「バディシステム」の構築です。
+
+このアプリケーションは3つの主要な問題を解決します：グループ活動の調整における面倒さ、一人で作業する際の先延ばし傾向、そして仲間の実際の努力や進捗が見えないという問題です。自動調整、リアルタイムエンゲージメント、活動の視覚的に認識することで、Buddy Callは仲間と一緒に取り組みたいあなたの行動を後押しします。
+
+## 要件
+
+### 要件1
+
+**ユーザーストーリー:** ユーザーとして、友人と勉強・目標グループを作成・管理したい。そうすることで活動を調整し、目標を共有できるから。
+
+#### 受入基準
+
+1. ユーザーが新しいグループを作成する時、システムはグループ名、共有目標の説明を設定し、招待リンクを生成できなければならない
+2. ユーザーが招待リンクを受け取った時、システムはリンクをクリックしてグループに参加できなければならない
+3. グループが作成された時、システムは最大2〜5人のメンバーをサポートしなければならない
+4. ユーザーがグループに所属している時、システムはグループ名、共有目標、全メンバーのリストを表示しなければならない
+5. ユーザーがグループを離れたい時、システムは退出を許可し、他のメンバーに通知しなければならない
+
+### 要件2
+
+**ユーザーストーリー:** グループメンバーとして、グループでコア集中時間をスケジュールしたい。そうすることで一緒に作業する時間を調整できるから。
+
+#### 受入基準
+
+1. グループメンバーがスケジュール機能にアクセスする時、システムは共有カレンダーインターフェースを表示しなければならない
+2. ユーザーが日時を選択する時、システムは最大60分までの集中セッションを提案できなければならない
+3. 集中セッションが提案された時、システムは全グループメンバーに確認通知を送らなければならない
+4. 全メンバーが確認した時、システムはセッションをスケジュールし、グループカレンダーに追加しなければならない
+5. スケジュールされたセッションが近づいた時、システムは全メンバーに通知を送る準備をしなければならない
+
+### 要件3
+
+**ユーザーストーリー:** グループメンバーとして、スケジュールされたセッションの自動通話風通知を受け取りたい。そうすることで覚えておいたり手動で他の人と調整する必要がないから。
+
+#### 受入基準
+
+1. スケジュールされた集中セッション時間になった時、システムは全グループメンバーに同時に電話風のプッシュ通知を送らなければならない
+2. ユーザーが通知を受け取った時、システムは「応答」と「辞退」のアクションオプションを提供しなければならない
+3. ユーザーが「応答」を選択した時、システムは自動的にアプリを起動し、集中ルームに遷移しなければならない
+4. ユーザーが「辞退」を選択した時、システムは理由スタンプ（例：「電車なう」「急用」「体調不良」）を表示しなければならない
+5. ユーザーが辞退理由を選択した時、システムはそのステータスを他のグループメンバーと共有しなければならない
+6. 通知に適切な時間内に応答がない時、システムはそれを辞退として扱わなければならない
+
+### 要件4
+
+**ユーザーストーリー:** 集中セッションの参加者として、活動のリアルタイムスナップショットを共有したい。そうすることでグループメンバーが私の進捗を見てモチベーションを維持できるから。
+
+#### 受入基準
+
+1. 集中セッションがアクティブな時、システムは最大60分までのセッション時間を追跡しなければならない
+2. 集中セッションタイマーが終了した時、システムは全アクティブ参加者に「スナップショットリクエスト」通知を送らなければならない
+3. ユーザーがスナップショットリクエストを受け取った時、システムは活動をキャプチャするための2分の制限時間を提供しなければならない
+4. ユーザーがスナップショットを撮る時、システムは写真撮影のためにリアカメラを自動的に起動しなければならない
+5. スナップショットがキャプチャされた時、システムは全参加者が見られる集中ルームタイムラインに投稿しなければならない
+6. ユーザーが制限時間内にスナップショットを撮らない場合、システムはそのセッションで無応答としてマークしなければならない
+
+### 要件5
+
+**ユーザーストーリー:** セッションを逃したグループメンバーとして、友人が何を達成したかを見たい。そうすることで参加できない時でもつながりとモチベーションを維持できるから。
+
+#### 受入基準
+
+1. 集中セッションがスナップショット付きで完了した時、システムは不在だったメンバーを特定しなければならない
+2. 完了したセッションからスナップショットが利用可能な時、システムは不在メンバーに通知を送らなければならない
+3. 不在メンバーに通知を送る時、システムは「友達が頑張っています！」のようなメッセージとセッションスナップショットを含めなければならない
+4. 不在メンバーが通知を受け取った時、システムはそのセッションの全スナップショットを表示できなければならない
+5. 不在メンバーがセッションスナップショットを見る時、システムは参加者に励ましやリアクションを送れなければならない
+
+### 要件6
+
+**ユーザーストーリー:** ユーザーとして、アプリがモバイルデバイスで確実に動作してほしい。そうすることで場所に関係なくセッションに参加できるから。
+
+#### 受入基準
+
+1. アプリがインストールされた時、システムはカメラ、通知、ストレージの必要な権限を要求しなければならない
+2. アプリが実行される時、システムは過去のセッションとグループ情報の表示でオフライン動作しなければならない
+3. ネットワーク接続が利用可能な時、システムは全データと通知を同期しなければならない
+4. アプリがバックグラウンドにある時、システムは依然としてプッシュ通知を受信・表示しなければならない
+5. スナップショットを撮る時、システムはカメラ権限を処理し、拒否された場合は適切なエラーメッセージを提供しなければならない
+
+### 要件7
+
+**ユーザーストーリー:** ユーザーとして、データと写真が安全でプライベートであってほしい。そうすることで個人情報をアプリに安心して託せるから。
+
+#### 受入基準
+
+1. ユーザーがアプリに参加する時、システムは認証（メール、電話、またはソーシャルログイン）を要求しなければならない
+2. 写真がキャプチャされた時、システムは特定のグループコンテキスト内でのみ共有しなければならない
+3. ユーザーがグループを離れた時、システムはそのグループのコンテンツへのアクセスを削除しなければならない
+4. データが保存される時、システムは機密情報を暗号化し、プライバシーのベストプラクティスに従わなければならない
+5. ユーザーがアカウントを削除する時、システムは残りメンバーのグループ機能を保持しながら全個人データを削除しなければならない
