+++ buddycall/.kiro/specs/buddy-call/tasks.md
@@ -1,0 +1,136 @@
+# 実装計画（2週間MVP版）
+
+## 第1週：基盤とコア機能
+
+- [ ] 1. 基本的なデータモデルとタイプ定義の実装
+  - TypeScriptインターフェースとenumの定義
+  - Firestoreデータ構造に対応したタイプ定義
+  - _要件: 1.1, 2.1, 3.1, 4.1, 5.1, 6.1, 7.1_
+
+- [ ] 2. Firebase設定とセキュリティルールの実装
+  - 基本的なFirestoreセキュリティルールの作成
+  - Firebase Storageセキュリティルールの設定
+  - _要件: 7.1, 7.2, 7.3_
+
+- [ ] 3. 基本的なユーティリティ関数とサービスの実装
+- [ ] 3.1 Firestoreデータアクセス層の実装
+  - 基本的なCRUD操作のためのサービス関数作成
+  - リアルタイムリスナーのセットアップ
+  - 簡単なエラーハンドリング機能の実装
+  - _要件: 1.1, 2.1, 6.3_
+
+- [ ] 3.2 認証サービスの拡張
+  - 既存の認証機能の活用
+  - ユーザープロフィール基本管理機能の追加
+  - _要件: 7.1_
+
+- [ ] 4. グループ管理機能の実装（簡易版）
+- [ ] 4.1 グループ作成機能の実装
+  - シンプルなグループ作成画面のUI実装
+  - Firestoreへのグループデータ保存
+  - 基本的な招待リンク生成機能の実装
+  - _要件: 1.1, 1.2_
+
+- [ ] 4.2 グループ参加機能の実装
+  - 招待リンク処理の基本実装
+  - グループ参加ロジックの実装
+  - _要件: 1.2, 1.3_
+
+- [ ] 4.3 グループ一覧表示機能の実装
+  - シンプルなグループ一覧表示
+  - 基本的なグループ詳細表示
+  - _要件: 1.4_
+
+## 第2週：セッション機能とMVP完成
+
+- [ ] 5. 簡易スケジュール管理機能の実装
+- [ ] 5.1 基本的なセッション作成機能
+  - シンプルな日時選択機能
+  - セッション作成とFirestore保存
+  - _要件: 2.1, 2.2_
+
+- [ ] 5.2 セッション一覧と参加確認機能
+  - 予定されたセッションの表示
+  - 参加・不参加の選択機能
+  - _要件: 2.3, 2.4_
+
+- [ ] 6. 基本的な通知機能の実装
+- [ ] 6.1 プッシュ通知の基本設定
+  - FCM基本設定とトークン管理
+  - 通知権限要求の実装
+  - _要件: 3.1, 6.1_
+
+- [ ] 6.2 セッション開始通知の実装
+  - 基本的なプッシュ通知送信
+  - 通知からアプリ起動への対応
+  - _要件: 3.1, 3.2, 3.3_
+
+- [ ] 7. 集中セッション機能の実装（簡易版）
+- [ ] 7.1 セッションルーム画面の実装
+  - 基本的なセッション画面
+  - 参加者表示機能
+  - シンプルなタイマー機能
+  - _要件: 4.1, 3.3_
+
+- [ ] 7.2 セッションタイマーの実装
+  - 60分最大タイマーの実装
+  - セッション開始・終了処理
+  - _要件: 4.1, 4.2_
+
+- [ ] 8. スナップショット機能の実装（簡易版）
+- [ ] 8.1 基本的なカメラ機能の実装
+  - Expo Cameraの統合
+  - 写真撮影機能の実装
+  - _要件: 4.4, 6.1_
+
+- [ ] 8.2 スナップショット共有の基本実装
+  - 撮影した写真のFirebase Storage保存
+  - セッション内での写真表示
+  - _要件: 4.5, 7.2_
+
+- [ ] 9. ホーム画面とナビゲーションの実装
+- [ ] 9.1 メインホーム画面の実装
+  - グループ一覧表示
+  - 次回セッション表示
+  - _要件: 1.4, 2.1_
+
+- [ ] 9.2 基本的なナビゲーションの実装
+  - タブナビゲーション設定
+  - 主要画面間の遷移
+  - _要件: 6.2_
+
+- [ ] 10. 基本的なテストの実装
+- [ ] 10.1 重要機能のユニットテスト
+  - データアクセス層のテスト
+  - 認証機能のテスト
+  - _要件: 全要件の品質保証_
+
+## 2週間後に延期する機能
+
+**高度な通知機能:**
+- 通話風フルスクリーン通知UI
+- 辞退理由スタンプ機能
+- 不在メンバーへの自動通知
+
+**高度なスナップショット機能:**
+- 自動スナップショット要求
+- 2分制限時間機能
+- タイムライン表示
+
+**追加機能:**
+- オフライン対応
+- 高度なエラーハンドリング
+- 包括的なテスト
+- プロフィール管理画面
+
+## MVP完成時の機能
+
+2週間で以下の基本機能が動作するMVPが完成します：
+
+1. ✅ ユーザー認証とログイン
+2. ✅ グループ作成と招待リンクでの参加
+3. ✅ セッションのスケジューリング
+4. ✅ 基本的なプッシュ通知
+5. ✅ セッションルームでのタイマー機能
+6. ✅ 写真撮影と共有機能
+7. ✅ シンプルなナビゲーション
